fix(pipeline): stabilize pipeline execution, redirections, and cleanup
fix(pipeline): clean up fork failure handling and move redirections into pipeline execution
fix(pipeline): enable redirections on single command
fix(pipeline): free pipeline commands and handle empty or trailing segments
fix(tokenizer): stop word tokens at pipe char
fix(types): use size_t for arg and token indexing
chore(execute): doc pipeline VLA tradeoffs and exit-status TODOs

+++ src/argvec.c
@@ -4,7 +4,7 @@
 #include "argvec.h"
 
 void double_argvec_capacity(ArgVec* argv) {
-    int new_cap = argv->cap ? argv->cap * 2 : 8;
+    size_t new_cap = argv->cap ? argv->cap * 2 : 8;
     char** new_args = realloc(argv->args, sizeof(char*) * new_cap);
     if(!new_args) {
         abort(); // Handle memory allocation failure
